Object_Detection_using_YOLOv3/ai-service-image/requirements.txt
Flask>=2.2.2
waitress==2.1.2
keras==2.10.0
tensorflow==2.10.0
pillow==8.4.0
matplotlib==3.4.3
transformers==4.21.0
numpy~=1.23
pandas==1.3.4
pillow==8.4.0